renamed back to suggest-list since diff isn't picked up anyways
