+++ search-ui/src/components/search/index.ts
@@ -1,4 +1,4 @@
 
 export { default as SearchBar } from './SearchBar.vue'
 export { default as SearchResults } from './SearchResults.vue'
-export { default as AutoComplete } from './AutoComplete.vue'
+export { default as SuggestList } from './SuggestList.vue'
